refactor: 인터페이스를 활용한 프록시 구현
