+++ app/src/main/java/com/techcourse/service/TransactionUserService.java
@@ -1,0 +1,59 @@
+package com.techcourse.service;
+
+import com.techcourse.domain.User;
+import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.datasource.DataSourceUtils;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public class TransactionUserService implements UserService {
+
+    private AppUserService appUserService;
+    private DataSource dataSource;
+
+
+    public TransactionUserService(final AppUserService appUserService, final DataSource dataSource) {
+        this.appUserService = appUserService;
+        this.dataSource = dataSource;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return appUserService.findById(id);
+    }
+
+    @Override
+    public void insert(final User user) {
+        appUserService.insert(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        Connection connection = null;
+        try {
+            connection = DataSourceUtils.getConnection(dataSource);
+            connection.setAutoCommit(false);
+
+            appUserService.changePassword(id, newPassword, createBy);
+
+            connection.commit();
+        } catch (SQLException | RuntimeException e) {
+            tryRollback(connection);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(dataSource);
+        }
+    }
+
+    private void tryRollback(Connection connection) {
+        try {
+            if (connection != null) {
+                connection.rollback();
+            }
+        } catch (SQLException ex) {
+            throw new DataAccessException(ex);
+        }
+    }
+}
